Split symlinking from installer into symlinker
