+++ src/cli/commands/switch.rs
@@ -4,7 +4,7 @@ use crate::{
     cli::{commands::HandleCommand, display::logging::warning},
     config::Config,
     error_handling::HandleError,
-    installer::{types::Version, Installer, InstallerOptions},
+    installer::{types::Version, Symlinker},
     repositories::manager::RepositoryManager,
     storage::package_register::PackageRegister,
 };
@@ -23,7 +23,7 @@ pub struct SwitchArgs {
 }
 
 impl HandleCommand for SwitchArgs {
-    fn handle(&self, config: &Config, manager: &RepositoryManager) {
+    fn handle(&self, config: &Config, _: &RepositoryManager) {
         let register_path = PackageRegister::get_default_path();
         let mut register = PackageRegister::from(&register_path).unwrap_or_exit(1);
 
@@ -53,8 +53,9 @@ impl HandleCommand for SwitchArgs {
         let should_symlink = !self.skip_symlinking && package.symlinked;
 
         // Set package version to active
-        let mut installer = Installer::new(&config, &mut register, &manager, InstallerOptions::default());
-        installer.set_active(&package_id, should_symlink).unwrap_or_exit_msg("Cannot switch active package", 1);
+        Symlinker::new(config)
+            .set_active(&mut register, &package_id, should_symlink)
+            .unwrap_or_exit_msg("Cannot switch active package", 1);
 
         // Save package register
         register.save_to(&register_path).unwrap_or_exit(1);
