+++ src/cli/commands/unlink.rs
@@ -1,12 +1,8 @@
 use clap::Args;
 
 use crate::{
-    cli::commands::HandleCommand,
-    config::Config,
-    error_handling::HandleError,
-    installer::{Installer, InstallerOptions},
-    repositories::manager::RepositoryManager,
-    storage::package_register::PackageRegister,
+    cli::commands::HandleCommand, config::Config, error_handling::HandleError, installer::Symlinker,
+    repositories::manager::RepositoryManager, storage::package_register::PackageRegister,
 };
 
 #[derive(Args, Debug)]
@@ -16,7 +12,7 @@ pub struct UnlinkArgs {
 }
 
 impl HandleCommand for UnlinkArgs {
-    fn handle(&self, config: &Config, manager: &RepositoryManager) {
+    fn handle(&self, config: &Config, _: &RepositoryManager) {
         let register_path = PackageRegister::get_default_path();
         let mut register = PackageRegister::from(&register_path).unwrap_or_exit(1);
 
@@ -32,8 +28,8 @@ impl HandleCommand for UnlinkArgs {
         }
 
         // Unlink package
-        Installer::new(config, &mut register, manager, InstallerOptions::default())
-            .unlink_package(&self.package_name)
+        Symlinker::new(config)
+            .unlink_package(&mut register, &self.package_name)
             .unwrap_or_exit_msg("Unable to unlink package", 1);
 
         // Save package register
