+++ src/installer/mod.rs
@@ -4,9 +4,12 @@ pub mod error;
 mod installer;
 mod options;
 pub mod scripts;
+mod symlinker;
 pub mod types;
 mod unpack;
 
 pub use self::installer::Installer;
 
 pub use self::options::InstallerOptions;
+
+pub use self::symlinker::Symlinker;
